Created a FuturesDataProvider interface
