+++ qf_lib/data_providers/abstract_price_data_provider.py
@@ -20,7 +20,6 @@ from numpy import nan
 from pandas._libs.tslibs.offsets import to_offset
 from pandas._libs.tslibs.timestamps import Timestamp
 
-from qf_lib.common.enums.expiration_date_field import ExpirationDateField
 from qf_lib.common.enums.frequency import Frequency
 from qf_lib.common.enums.price_field import PriceField
 from qf_lib.common.tickers.tickers import Ticker
@@ -28,7 +27,6 @@ from qf_lib.common.utils.dateutils.relative_delta import RelativeDelta
 from qf_lib.common.utils.miscellaneous.to_list_conversion import convert_to_list
 from qf_lib.containers.dataframe.prices_dataframe import PricesDataFrame
 from qf_lib.containers.dataframe.qf_dataframe import QFDataFrame
-from qf_lib.containers.futures.future_tickers.future_ticker import FutureTicker
 from qf_lib.containers.qf_data_array import QFDataArray
 from qf_lib.containers.series.prices_series import PricesSeries
 from qf_lib.containers.series.qf_series import QFSeries
@@ -99,7 +97,8 @@ class AbstractPriceDataProvider(DataProvider, metaclass=ABCMeta):
 
         # Map the specific fields onto the fields given by the str_to_field_dict
         if isinstance(container, QFDataArray):
-            container = container.assign_coords(fields=[str_to_field_dict[field_str] for field_str in container.fields.values])
+            container = container.assign_coords(
+                fields=[str_to_field_dict[field_str] for field_str in container.fields.values])
             normalized_result = normalize_data_array(
                 container, tickers, fields, got_single_date, got_single_ticker, got_single_field, use_prices_types=True
             )
@@ -113,7 +112,7 @@ class AbstractPriceDataProvider(DataProvider, metaclass=ABCMeta):
         return normalized_result
 
     @abstractmethod
-    def price_field_to_str_map(self) -> Dict[PriceField, str]:
+    def price_field_to_str_map(self, *args) -> Dict[PriceField, str]:
         """
         Method has to be implemented in each data provider in order to be able to use get_price.
         Returns dictionary containing mapping between PriceField and corresponding string that has to be used by
@@ -125,65 +124,6 @@ class AbstractPriceDataProvider(DataProvider, metaclass=ABCMeta):
              mapping between PriceFields and corresponding strings
         """
         raise NotImplementedError()
-
-    @abstractmethod
-    def expiration_date_field_str_map(self, ticker: Ticker = None) -> Dict[ExpirationDateField, str]:
-        """
-        Method has to be implemented in each data provider in order to be able to use get_futures_chain_tickers.
-        Returns dictionary containing mapping between ExpirationDateField and corresponding string that has to be used
-        by get_futures_chain_tickers method.
-
-        Parameters
-        -----------
-        ticker: None, Ticker
-            ticker is optional and might be uses by particular data providers to create appropriate dictionary
-
-        Returns
-        -------
-        Dict[ExpirationDateField, str]
-             mapping between ExpirationDateField and corresponding strings
-        """
-        pass
-
-    def get_futures_chain_tickers(self, tickers: Union[FutureTicker, Sequence[FutureTicker]],
-                                  expiration_date_fields: Union[ExpirationDateField, Sequence[ExpirationDateField]]) \
-            -> Dict[FutureTicker, QFDataFrame]:
-        """
-        Returns tickers of futures contracts, which belong to the same futures contract chain as the provided ticker
-        (tickers), along with their expiration dates in form of a QFSeries or QFDataFrame.
-
-        Parameters
-        ----------
-        tickers: FutureTicker, Sequence[FutureTicker]
-            tickers for which should the future chain tickers be retrieved
-        expiration_date_fields: ExpirationDateField, Sequence[ExpirationDateField]
-            field that should be downloaded as the expiration date field, by default last tradeable date
-
-        Returns
-        -------
-        Dict[FutureTicker, QFDataFrame]
-            Returns a dictionary, which maps Tickers to QFSeries, consisting of the expiration dates of Future
-            Contracts: Dict[FutureTicker, QFDataFrame]]. The QFDataFrames contain the
-            specific Tickers, which belong to the corresponding futures family, same as the FutureTicker, and are
-            indexed by the expiration dates of the specific future contracts.
-        """
-        expiration_date_fields, got_single_expiration_date_field = convert_to_list(expiration_date_fields,
-                                                                                   ExpirationDateField)
-        mapping_dict = self.expiration_date_field_str_map()
-        expiration_date_fields_str = [mapping_dict[field] for field in expiration_date_fields]
-        exp_dates_dict = self._get_futures_chain_dict(tickers, expiration_date_fields_str)
-
-        for future_ticker, exp_dates in exp_dates_dict.items():
-            exp_dates = exp_dates.rename(columns=self.str_to_expiration_date_field_map())
-            for ticker in exp_dates.index:
-                ticker.security_type = future_ticker.security_type
-                ticker.point_value = future_ticker.point_value
-                ticker.set_name(future_ticker.name)
-            if got_single_expiration_date_field:
-                exp_dates = exp_dates.squeeze()
-            exp_dates_dict[future_ticker] = exp_dates
-
-        return exp_dates_dict
 
     def historical_price(self, tickers: Union[Ticker, Sequence[Ticker]],
                          fields: Union[PriceField, Sequence[PriceField]],
@@ -287,8 +227,17 @@ class AbstractPriceDataProvider(DataProvider, metaclass=ABCMeta):
         open_prices = self.get_price(tickers, PriceField.Open, start_date, end_time, frequency)
         close_prices = self.get_price(tickers, PriceField.Close, start_date, end_time, frequency)
 
-        latest_available_prices_series = self._get_valid_latest_available_prices(start_date, tickers, open_prices, close_prices)
+        latest_available_prices_series = self._get_valid_latest_available_prices(start_date, tickers, open_prices,
+                                                                                 close_prices)
         return latest_available_prices_series.iloc[0] if got_single_ticker else latest_available_prices_series
+
+    def str_to_price_field_map(self) -> Dict[str, PriceField]:
+        """
+        Inverse of price_field_to_str_map.
+        """
+        field_str_dict = self.price_field_to_str_map()
+        inv_dict = {v: k for k, v in field_str_dict.items()}
+        return inv_dict
 
     @staticmethod
     def _get_valid_latest_available_prices(start_date: datetime, tickers: Sequence[Ticker], open_prices: QFDataFrame,
@@ -342,41 +291,7 @@ class AbstractPriceDataProvider(DataProvider, metaclass=ABCMeta):
     @staticmethod
     def _got_single_date(start_date: datetime, end_date: datetime, frequency: Frequency):
         return start_date.date() == end_date.date() if frequency <= Frequency.DAILY else \
-                (start_date + frequency.time_delta() > end_date)
-
-    @abstractmethod
-    def _get_futures_chain_dict(self, tickers: Union[FutureTicker, Sequence[FutureTicker]],
-                                expiration_date_fields: Union[str, Sequence[str]]) -> Dict[FutureTicker, QFDataFrame]:
-        """
-        Returns a dictionary, which maps Tickers to QFSeries, consisting of the expiration dates of Future
-        Contracts: Dict[FutureTicker, Union[QFSeries, QFDataFrame]]].
-
-        Parameters
-        ----------
-        tickers: Ticker, Sequence[Ticker]
-            tickers for securities which should be retrieved
-        expiration_date_fields: str, Sequence[str]
-            expiration date fields of securities which should be retrieved. Specific for each data provider,
-            the mapping between strings and corresponding ExpirationDateField enum values should be implemented as
-            str_to_expiration_date_field_map function.
-        """
-        pass
-
-    def str_to_expiration_date_field_map(self, ticker: Ticker = None) -> Dict[str, ExpirationDateField]:
-        """
-        Inverse of str_to_expiration_date_field_map.
-        """
-        field_str_dict = self.expiration_date_field_str_map(ticker)
-        inv_dict = {v: k for k, v in field_str_dict.items()}
-        return inv_dict
-
-    def str_to_price_field_map(self) -> Dict[str, PriceField]:
-        """
-        Inverse of price_field_to_str_map.
-        """
-        field_str_dict = self.price_field_to_str_map()
-        inv_dict = {v: k for k, v in field_str_dict.items()}
-        return inv_dict
+            (start_date + frequency.time_delta() > end_date)
 
     def _map_field_to_str(
             self, fields: Union[None, PriceField, Sequence[PriceField]]) \
